refactor(nxext): remove serve builder and use build with watch/serve as default
