+++ packages/stencil/migrations.json
@@ -14,6 +14,11 @@
       "version": "0.1.3",
       "description": "update-0-1-3",
       "factory": "./src/migrations/update-0-1-3/update-0-1-3"
+    },
+    "update-1-0-0-beta0": {
+      "version": "1.0.0.beta0",
+      "description": "update-1-0-0-beta0",
+      "factory": "./src/migrations/update-1-0-0-beta0/update-1-0-0-beta0"
     }
   }
 }
