--- packages/stencil/src/builders/serve/schema.json
@@ -1,41 +1,0 @@
-{
-  "$schema": "https://json-schema.org/draft-07/schema",
-  "$id": "https://json-schema.org/draft-07/schema",
-  "title": "Serve builder",
-  "description": "",
-  "type": "object",
-  "properties": {
-    "projectType": {
-      "type": "string"
-    },
-    "configPath": {
-      "type": "string"
-    },
-    "debug": {
-      "type": "boolean"
-    },
-    "dev": {
-      "type": "boolean"
-    },
-    "docs": {
-      "type": "boolean"
-    },
-    "es5": {
-      "type": "boolean"
-    },
-    "port": {
-      "type": "number"
-    },
-    "verbose": {
-      "type": "boolean"
-    },
-    "watch": {
-      "type": "boolean"
-    },
-    "open": {
-      "type": "boolean",
-      "default": true
-    }
-  },
-  "required": []
-}
